Optimizing Michigan

+++ catscripts/Government/Countries/United_States/US_states/Michigan/Michigan.sh
@@ -5,25 +5,40 @@ KEYWORDS_DETROIT="Detroit"
 KEYWORDS_MICHIGAN_EXCLUDE="$KEYWORDS_DETROIT|Lake(| )Michigan"
 KEYWORDS_DETROIT_EXCLUDE="Detroit(| )Lions"
 
-egrep -i "$KEYWORDS_MICHIGAN" newpages.txt | egrep -i "$KEYWORDS_MiCHIGAN_EXCLUDE" >> Michigan.txt
-egrep -i "$KEYWORDS_DETROIT" newpages.txt | egrep -iv "$KEYWORDS_DETROIT_EXCLUDE" >> Detroit.txt
-
-MICHIGAN=`stat --print=%s Michigan.txt`
-DETROIT=`stat --print=%s Detroit.txt`
-
-if [ $MICHIGAN -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="Michigan.txt"
-  export CATNAME="Michigan"
-  $CATEGORIZE
-fi
 
-if [ $DETROIT -ne 0 ];
-then
-  export CATFILE="Detroit.txt"
-  export CATNAME="Detroit"
-  $CATEGORIZE
-fi
+  if [ "$DEBUG" == "yes" ];
+  then
+    printf "Starting Michigan\n"
+  fi
 
-rm Michigan.txt
-rm Detroit.txt
+  MICHIGAN=`egrep -i "$KEYWORDS_MICHIGAN" newpages.txt | egrep -iv "$KEYWORDS_MICHIGAN_EXCLUDE"`
+  DETROIT=`egrep -i "$KEYWORDS_DETROIT" newpages.txt | egrep -iv "$KEYWORDS_DETROIT_EXCLUDE"`
+
+  if [ "$MICHIGAN" != "" ];
+  then
+    printf "$MICHIGAN" > Michigan.txt
+    export CATFILE="Michigan.txt"
+    export CATNAME="Michigan"
+    $CATEGORIZE
+    rm Michigan.txt
+    unset MICHIGAN
+  fi
+
+  if [ "$DETROIT" != "" ];
+  then
+    printf "$DETROIT" > Detroit.txt
+    export CATFILE="Detroit.txt"
+    export CATNAME="Detroit"
+    $CATEGORIZE
+    rm Detroit.txt
+    unset DETROIT
+  fi
+
+  if [ "$DEBUG" == "yes" ];
+  then
+    printf "Finishing Michigan\n"
+  fi
+
+fi
